requirements.txt
datasets
huggingface_hub
python-dotenv==1.0.1
tiktoken
tqdm